[added] Feature manager for server customization. XP scaling and damage modifiers customizable. [updated] XP scaling to work with all xp txpes. [added] Pet pvp damage reduction.
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@1502 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/zone/managers/player/PlayerManager.h
@@ -91,9 +91,9 @@ public:
 
 	PlayerMap* getPlayerMap();
 
-	void setXpScale(int scale);
+	void setXpScale(float scale);
 
-	int getXpScale();
+	float getXpScale();
 
 protected:
 	PlayerManager(DummyConstructorParameter* param);
@@ -175,9 +175,9 @@ public:
 
 	PlayerMap* getPlayerMap();
 
-	void setXpScale(int scale);
+	void setXpScale(float scale);
 
-	int getXpScale();
+	float getXpScale();
 
 protected:
 	String _param0_validateName__String_;
